Move: to a single cli app
