+++ src/taker/mod.rs
@@ -10,8 +10,6 @@ pub mod error;
 pub mod offers;
 mod routines;
 
-pub mod rpc;
-
 pub use self::api::TakerBehavior;
 pub use api::{SwapParams, Taker};
 pub use config::TakerConfig;
