--- src/taker/rpc/mod.rs
@@ -1,5 +1,0 @@
-mod messages;
-mod server;
-
-pub use messages::{RpcMsgReq, RpcMsgResp};
-pub use server::start_taker_rpc_server;
